@@ -4,7 +4,7 @@ version: 0.0.1
 author: Ashvin <warrierar@gmail.com>
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   aqueduct: 3.3.0+1
