@@ -1,6 +1,8 @@
 import 'package:meta/meta.dart';
 import 'package:server/server.dart';
 
+import 'Field.dart' as fld;
+import 'Institution.dart';
 import 'User.dart';
 
 enum Degree {
@@ -8,6 +10,32 @@ enum Degree {
   bachelor,
   master,
   doctoral
+}
+
+String degreeToString(Degree degree) {
+  switch (degree) {
+    case Degree.associate:
+      return 'Associate';
+    case Degree.bachelor:
+      return 'Bachelor\'s';
+    case Degree.master:
+      return 'Master\'s';
+    default:
+      return 'Doctoral';
+  }
+}
+
+Degree stringToDegree(String string) {
+  switch (string) {
+    case 'Associate':
+      return Degree.associate;
+    case 'Bachelor\'s':
+      return Degree.bachelor;
+    case 'Master\'s':
+      return Degree.master;
+    default:
+      return Degree.doctoral;
+  }
 }
 
 class Education extends Serializable {
@@ -25,9 +53,9 @@ class Education extends Serializable {
 
   int id;
   User user;
-  String institution;
+  Institution institution;
   Degree degree;
-  String field;
+  fld.Field field;
   bool current;
   DateTime startDate;
   DateTime endDate;
@@ -35,14 +63,14 @@ class Education extends Serializable {
   @override
   Map<String, dynamic> asMap() {
     return {
-      "id": id,
-      "user": user.asMap(),
-      "institution": institution,
-      "degree": degree.toString(),
-      "field": field,
-      "current": current,
-      "startDate": startDate.toIso8601String(),
-      "endDate": endDate.toIso8601String()
+      'id': id,
+      'user': user.asMap(),
+      'institution': institution.asMap(),
+      'degree': degreeToString(degree),
+      'field': field.asMap(),
+      'current': current,
+      'startDate': startDate.toIso8601String(),
+      'endDate': endDate.toIso8601String()
     };
   }
 
@@ -50,7 +78,7 @@ class Education extends Serializable {
   void readFromMap(Map<String, dynamic> object) {
     if (user == null) {
       final userMap = object['user'] as Map<String, dynamic>;
-      if (userMap['type'] == UserType.student.toString()) {
+      if (stringToUserType(userMap['type'] as String) == UserType.student) {
         user = Student()
           ..readFromMap(userMap);
       } else {
@@ -58,13 +86,52 @@ class Education extends Serializable {
           ..readFromMap(userMap);
       }
     }
-    institution = object['institution'] as String;
-    final degreeString = object['degree'] as String;
-    degree = Degree.values
-        .firstWhere((element) => degreeString == element.toString());
-    field = object['field'] as String;
+    institution = Institution()
+      ..readFromMap(object['institution'] as Map<String, dynamic>);
+    degree = stringToDegree(object['degree'] as String);
+    field = fld.Field()
+      ..readFromMap(object['field'] as Map<String, dynamic>);
     current = object['current'] as bool;
     startDate = DateTime.parse(object['startDate'] as String);
     endDate = DateTime.parse(object['endDate'] as String);
+  }
+
+  Future<void> save() async {
+    const sql = '''
+      INSERT INTO education
+      (user, institution, degree, field, current, start_date, end_date)
+      VALUES (?, ?, ?, ?, ?, ?, ?)
+    ''';
+    await ServerChannel.db.query(sql, [
+      user.id,
+      institution.name,
+      degreeToString(degree),
+      field.name,
+      current,
+      startDate.toUtc(),
+      endDate.toUtc()
+    ]);
+  }
+
+  static Future<List<Education>> getByUser(User user) async {
+    const sql = '''
+      SELECT * FROM education
+      WHERE user = ?
+    ''';
+    final results = await ServerChannel.db.query(sql, [user.id]);
+
+    final resultFutures = results.map((e) async =>
+        Education.create(
+            user: user,
+            institution: await Institution.get(e[2] as String),
+            degree: stringToDegree(e[3] as String),
+            field: fld.Field.create(name: e[4] as String),
+            current: (e[5] as int) == 1,
+            startDate: (e[6] as DateTime).toLocal(),
+            endDate: (e[7] as DateTime)?.toLocal()
+        )
+          ..id = e[0] as int
+    );
+    return Future.wait(resultFutures);
   }
 }
