@@ -1,12 +1,7 @@
 import 'package:meta/meta.dart';
 
 import '../../server.dart';
-import '../db/Education.dart';
-import '../db/Interest.dart';
-import '../db/Skill.dart';
-import '../db/User.dart';
-import '../db/Volunteering.dart';
-import '../db/Work.dart';
+import '../db/models.dart';
 
 class CardInfo extends Serializable {
   CardInfo({
@@ -29,15 +24,21 @@ class CardInfo extends Serializable {
   Map<String, dynamic> asMap() {
     return {
       "user": user.asMap(),
-      "education": education.map((e) => e.asMap())
+      "education": education.map((e) => e.asMap()
+        ..update('field', (value) => value['name'])
+        ..remove('user'))
           .toList(),
-      "work": work.map((e) => e.asMap())
+      "work": work.map((e) => e.asMap()
+        ..update('company', (value) => value['name'])
+        ..remove('user'))
           .toList(),
-      "volunteering": volunteering.map((e) => e.asMap())
+      "volunteering": volunteering.map((e) => e.asMap()
+        ..update('company', (value) => value['name'])
+        ..remove('user'))
           .toList(),
-      "skills": skills.map((e) => e.asMap())
+      "skills": skills.map((e) => e.title)
           .toList(),
-      "interests": interests.map((e) => e.asMap())
+      "interests": interests.map((e) => e.title)
           .toList()
     };
   }
@@ -75,5 +76,16 @@ class CardInfo extends Serializable {
     interests = interestsList.map((e) => Interest()
       ..readFromMap(e as Map<String, dynamic>))
         .toList();
+  }
+
+  static Future<CardInfo> get(User user) async {
+    return CardInfo(
+        user: user,
+        education: await Education.getByUser(user),
+        work: await Work.getByUser(user),
+        volunteering: await Volunteering.getByUser(user),
+        skills: await Skill.getByUser(user),
+        interests: await Interest.getByUser(user)
+    );
   }
 }
